Move ChainFactory out of experimental Add InternalVariable to Variable
